+++ src/calendar.vue
@@ -1,37 +1,22 @@
 <template>
-  <div class="mx-calendar">
+  <div class="mx-calendar" :class="'mx-calendar-panel-' + panel.toLowerCase()">
     <div class="mx-calendar-header">
-      <a
-        v-show="panel !== 'TIME'"
-        class="mx-icon-last-year"
-        @click="handleIconYear(-1)">&laquo;</a>
-      <a
-        v-show="panel === 'DATE'"
-        class="mx-icon-last-month"
-        @click="handleIconMonth(-1)">&lsaquo;</a>
-      <a
-        v-show="panel !== 'TIME'"
-        class="mx-icon-next-year"
-        @click="handleIconYear(1)">&raquo;</a>
-      <a
-        v-show="panel === 'DATE'"
-        class="mx-icon-next-month"
-        @click="handleIconMonth(1)">&rsaquo;</a>
+      <a v-show="panel !== 'TIME'" class="mx-icon-last-year" @click="handleIconYear(-1)">&#171;</a>
+      <a v-show="panel === 'DATE'" class="mx-icon-last-month" @click="handleIconMonth(-1)">&#8249;</a>
+      <a v-show="panel !== 'TIME'" class="mx-icon-next-year" @click="handleIconYear(1)">&#187;</a>
+      <a v-show="panel === 'DATE'" class="mx-icon-next-month" @click="handleIconMonth(1)">&#8250;</a>
       <a
         v-show="panel === 'DATE'"
         class="mx-current-month"
-        @click="handleBtnMonth">{{months[calendarMonth]}}</a>
+        @click="handleBtnMonth"
+      >{{months[calendarMonth]}}</a>
       <a
         v-show="panel === 'DATE' || panel === 'MONTH'"
         class="mx-current-year"
-        @click="handleBtnYear">{{calendarYear}}</a>
-      <a
-        v-show="panel === 'YEAR'"
-        class="mx-current-year">{{yearHeader}}</a>
-      <a
-        v-show="panel === 'TIME'"
-        class="mx-time-header"
-        @click="handleTimeHeader">{{timeHeader}}</a>
+        @click="handleBtnYear"
+      >{{calendarYear}}</a>
+      <a v-show="panel === 'YEAR'" class="mx-current-year">{{yearHeader}}</a>
+      <a v-show="panel === 'TIME'" class="mx-time-header" @click="handleTimeHeader">{{timeHeader}}</a>
     </div>
     <div class="mx-calendar-content">
       <panel-date
@@ -44,28 +29,33 @@
         :end-at="endAt"
         :first-day-of-week="firstDayOfWeek"
         :disabled-date="isDisabledDate"
-        @select="selectDate"/>
+        @select="selectDate"
+      />
       <panel-year
         v-show="panel === 'YEAR'"
         :value="value"
         :disabled-year="isDisabledYear"
         :first-year="firstYear"
-        @select="selectYear" />
+        @select="selectYear"
+      />
       <panel-month
         v-show="panel === 'MONTH'"
         :value="value"
         :disabled-month="isDisabledMonth"
         :calendar-year="calendarYear"
-        @select="selectMonth" />
+        @select="selectMonth"
+      />
       <panel-time
         v-show="panel === 'TIME'"
         :minute-step="minuteStep"
         :time-picker-options="timePickerOptions"
+        :time-select-options="timeSelectOptions"
         :value="value"
         :disabled-time="isDisabledTime"
         :time-type="timeType"
         @select="selectTime"
-        @pick="pickTime" />
+        @pick="pickTime"
+      />
     </div>
   </div>
 </template>
@@ -105,7 +95,13 @@ export default {
       type: String,
       default: 'YYYY-MM-DD'
     },
+    index: Number,
     // below user set
+    defaultValue: {
+      validator: function (val) {
+        return isValidDate(val)
+      }
+    },
     firstDayOfWeek: {
       default: 7,
       type: Number,
@@ -134,6 +130,12 @@ export default {
       default: 0,
       validator: val => val >= 0 && val <= 60
     },
+    timeSelectOptions: {
+      type: Object,
+      default () {
+        return null
+      }
+    },
     timePickerOptions: {
       type: [Object, Function],
       default () {
@@ -142,7 +144,7 @@ export default {
     }
   },
   data () {
-    const now = new Date()
+    const now = this.getNow(this.value)
     const calendarYear = now.getFullYear()
     const calendarMonth = now.getMonth()
     const firstYear = Math.floor(calendarYear / 10) * 10
@@ -177,7 +179,7 @@ export default {
       return this.value && formatDate(this.value, this.dateFormat)
     },
     yearHeader () {
-      return this.firstYear + ' ~ ' + (this.firstYear + 10)
+      return this.firstYear + ' ~ ' + (this.firstYear + 9)
     },
     months () {
       return this.t('months')
@@ -234,13 +236,22 @@ export default {
         this.updateNow(this.value)
       }
     },
+    getNow (value) {
+      return value
+        ? new Date(value)
+        : this.defaultValue && isValidDate(this.defaultValue)
+          ? new Date(this.defaultValue)
+          : new Date()
+    },
     // 根据value更新日历
     updateNow (value) {
-      const now = value ? new Date(value) : new Date()
-      const oldNow = new Date(this.now)
-      this.now = now
-      if (this.visible) {
-        this.dispatch('DatePicker', 'calendar-change', [now, oldNow])
+      const oldNow = this.now
+      this.now = this.getNow(value)
+      if (this.visible && this.now !== oldNow) {
+        this.dispatch('DatePicker', 'calendar-change', [
+          new Date(this.now),
+          new Date(oldNow)
+        ])
       }
     },
     getCriticalTime (value) {
@@ -258,14 +269,22 @@ export default {
       return date.getTime()
     },
     inBefore (time, startAt) {
-      startAt = startAt || this.startAt
-      return (this.notBeforeTime && time < this.notBeforeTime) ||
+      if (startAt === undefined) {
+        startAt = this.startAt
+      }
+      return (
+        (this.notBeforeTime && time < this.notBeforeTime) ||
         (startAt && time < this.getCriticalTime(startAt))
+      )
     },
     inAfter (time, endAt) {
-      endAt = endAt || this.endAt
-      return (this.notAfterTime && time > this.notAfterTime) ||
+      if (endAt === undefined) {
+        endAt = this.endAt
+      }
+      return (
+        (this.notAfterTime && time > this.notAfterTime) ||
         (endAt && time > this.getCriticalTime(endAt))
+      )
     },
     inDisabledDays (time) {
       if (Array.isArray(this.disabledDays)) {
@@ -278,21 +297,37 @@ export default {
     isDisabledYear (year) {
       const time = new Date(year, 0).getTime()
       const maxTime = new Date(year + 1, 0).getTime() - 1
-      return this.inBefore(maxTime) || this.inAfter(time) || (this.type === 'year' && this.inDisabledDays(time))
+      return (
+        this.inBefore(maxTime) ||
+        this.inAfter(time) ||
+        (this.type === 'year' && this.inDisabledDays(time))
+      )
     },
     isDisabledMonth (month) {
       const time = new Date(this.calendarYear, month).getTime()
       const maxTime = new Date(this.calendarYear, month + 1).getTime() - 1
-      return this.inBefore(maxTime) || this.inAfter(time) || (this.type === 'month' && this.inDisabledDays(time))
+      return (
+        this.inBefore(maxTime) ||
+        this.inAfter(time) ||
+        (this.type === 'month' && this.inDisabledDays(time))
+      )
     },
     isDisabledDate (date) {
       const time = new Date(date).getTime()
       const maxTime = new Date(date).setHours(23, 59, 59, 999)
-      return this.inBefore(maxTime) || this.inAfter(time) || this.inDisabledDays(time)
+      return (
+        this.inBefore(maxTime) ||
+        this.inAfter(time) ||
+        this.inDisabledDays(time)
+      )
     },
     isDisabledTime (date, startAt, endAt) {
       const time = new Date(date).getTime()
-      return this.inBefore(time, startAt) || this.inAfter(time, endAt) || this.inDisabledDays(time)
+      return (
+        this.inBefore(time, startAt) ||
+        this.inAfter(time, endAt) ||
+        this.inDisabledDays(time)
+      )
     },
     selectDate (date) {
       if (this.type === 'datetime') {
@@ -306,10 +341,16 @@ export default {
         }
         if (this.isDisabledTime(time)) {
           time.setHours(0, 0, 0, 0)
-          if (this.notBefore && time.getTime() < new Date(this.notBefore).getTime()) {
+          if (
+            this.notBefore &&
+            time.getTime() < new Date(this.notBefore).getTime()
+          ) {
             time = new Date(this.notBefore)
           }
-          if (this.startAt && time.getTime() < new Date(this.startAt).getTime()) {
+          if (
+            this.startAt &&
+            time.getTime() < new Date(this.startAt).getTime()
+          ) {
             time = new Date(this.startAt)
           }
         }
@@ -324,6 +365,7 @@ export default {
       if (this.type.toLowerCase() === 'year') {
         return this.selectDate(new Date(this.now))
       }
+      this.dispatch('DatePicker', 'select-year', [year, this.index])
       this.showPanelMonth()
     },
     selectMonth (month) {
@@ -331,6 +373,7 @@ export default {
       if (this.type.toLowerCase() === 'month') {
         return this.selectDate(new Date(this.now))
       }
+      this.dispatch('DatePicker', 'select-month', [month, this.index])
       this.showPanelDate()
     },
     selectTime (time) {
@@ -346,7 +389,9 @@ export default {
       this.updateNow(new Date(this.calendarYear, month))
     },
     getSibling () {
-      const calendars = this.$parent.$children.filter(v => v.$options.name === this.$options.name)
+      const calendars = this.$parent.$children.filter(
+        v => v.$options.name === this.$options.name
+      )
       const index = calendars.indexOf(this)
       const sibling = calendars[index ^ 1]
       return sibling