+++ CHANGELOG.md
@@ -1,12 +1,219 @@
-<a name="2.5.0"></a>
+# [2.12.0](https://github.com/mengxiong10/vue2-datepicker/compare/v2.11.2...v2.12.0) (2019-06-25)
+
+
+### Features
+
+* add the prop `time-select-options` ([#227](https://github.com/mengxiong10/vue2-datepicker/issues/227)) ([a55b4b6](https://github.com/mengxiong10/vue2-datepicker/commit/a55b4b6))
+
+
+
+## [2.11.2](https://github.com/mengxiong10/vue2-datepicker/compare/v2.11.1...v2.11.2) (2019-05-15)
+
+
+### Bug Fixes
+
+* fecha.parse invalid return false ([#309](https://github.com/mengxiong10/vue2-datepicker/issues/309)) ([96812ab](https://github.com/mengxiong10/vue2-datepicker/commit/96812ab))
+* remove last shortcut "|" ([c6a6300](https://github.com/mengxiong10/vue2-datepicker/commit/c6a6300))
+* test suites in another timezone failed ([#300](https://github.com/mengxiong10/vue2-datepicker/issues/300)) ([ec69590](https://github.com/mengxiong10/vue2-datepicker/commit/ec69590))
+
+
+
+## [2.11.1](https://github.com/mengxiong10/vue2-datepicker/compare/v2.11.0...v2.11.1) (2019-04-30)
+
+
+### Bug Fixes
+
+* close popup after select range ([5823f85](https://github.com/mengxiong10/vue2-datepicker/commit/5823f85))
+
+
+
+# [2.11.0](https://github.com/mengxiong10/vue2-datepicker/compare/v2.10.3...v2.11.0) (2019-04-09)
+
+
+### Bug Fixes
+
+* clickoutside bad when use append-to-body ([#291](https://github.com/mengxiong10/vue2-datepicker/issues/291)) ([9bb6046](https://github.com/mengxiong10/vue2-datepicker/commit/9bb6046))
+
+
+### Features
+
+*  `select-year` and `select-month`  ([#290](https://github.com/mengxiong10/vue2-datepicker/issues/290)) ([20d1f19](https://github.com/mengxiong10/vue2-datepicker/commit/20d1f19))
+
+
+
+## [2.10.3](https://github.com/mengxiong10/vue2-datepicker/compare/v2.10.2...v2.10.3) (2019-03-14)
+
+
+### Bug Fixes
+
+* type input don't change the value ([#266](https://github.com/mengxiong10/vue2-datepicker/issues/266)) ([aa20a1e](https://github.com/mengxiong10/vue2-datepicker/commit/aa20a1e))
+
+
+
+## [2.10.2](https://github.com/mengxiong10/vue2-datepicker/compare/v2.10.1...v2.10.2) (2019-03-14)
+
+
+### Bug Fixes
+
+* pick minute will affect the hour ([#168](https://github.com/mengxiong10/vue2-datepicker/issues/168)) ([2afed88](https://github.com/mengxiong10/vue2-datepicker/commit/2afed88))
+* remove the append background ([e7775d6](https://github.com/mengxiong10/vue2-datepicker/commit/e7775d6))
+
+
+### Performance Improvements
+
+* remove IE 10's clear button X ([6a990d8](https://github.com/mengxiong10/vue2-datepicker/commit/6a990d8))
+
+
+
+## [2.10.1](https://github.com/mengxiong10/vue2-datepicker/compare/v2.10.0...v2.10.1) (2019-03-11)
+
+
+### Performance Improvements
+
+* change the entity name to entity number ([dbbba6b](https://github.com/mengxiong10/vue2-datepicker/commit/dbbba6b))
+
+
+
+# [2.10.0](https://github.com/mengxiong10/vue2-datepicker/compare/v2.9.2...v2.10.0) (2019-02-12)
+
+
+### Features
+
+* add prop `default-value` for calendar default date ([#94](https://github.com/mengxiong10/vue2-datepicker/issues/94)) ([4ff6945](https://github.com/mengxiong10/vue2-datepicker/commit/4ff6945))
+* show the popup on focus and close it on blur ([3bcedf5](https://github.com/mengxiong10/vue2-datepicker/commit/3bcedf5))
+
+
+
+## [2.9.2](https://github.com/mengxiong10/vue2-datepicker/compare/v2.9.1...v2.9.2) (2019-02-10)
+
+
+### Bug Fixes
+
+* the display error of the year panel header ([#245](https://github.com/mengxiong10/vue2-datepicker/issues/245)) ([7bc2785](https://github.com/mengxiong10/vue2-datepicker/commit/7bc2785))
+
+
+
+## [2.9.1](https://github.com/mengxiong10/vue2-datepicker/compare/v2.9.0...v2.9.1) (2019-02-01)
+
+
+### Bug Fixes
+
+* stringify function returns a error value,when date is null ([#244](https://github.com/mengxiong10/vue2-datepicker/issues/244)) ([92243ab](https://github.com/mengxiong10/vue2-datepicker/commit/92243ab))
+
+
+
+# [2.9.0](https://github.com/mengxiong10/vue2-datepicker/compare/v2.8.1...v2.9.0) (2019-01-29)
+
+
+### Features
+
+* support custom format function ([c801516](https://github.com/mengxiong10/vue2-datepicker/commit/c801516))
+
+
+
+## [2.8.1](https://github.com/mengxiong10/vue2-datepicker/compare/v2.8.0...v2.8.1) (2019-01-24)
+
+
+### Bug Fixes
+
+* clickoutside event listener change capturing to bubbling ([054758e](https://github.com/mengxiong10/vue2-datepicker/commit/054758e))
+* use refs instead of children array ([f43e3a3](https://github.com/mengxiong10/vue2-datepicker/commit/f43e3a3))
+
+
+
+# [2.8.0](https://github.com/mengxiong10/vue2-datepicker/compare/v2.7.0...v2.8.0) (2019-01-13)
+
+
+### Features
+
+* add `valueType` to format binding value ([dd6f2ea](https://github.com/mengxiong10/vue2-datepicker/commit/dd6f2ea))
+
+
+
+# [2.7.0](https://github.com/mengxiong10/vue2-datepicker/compare/v2.6.3...v2.7.0) (2019-01-08)
+
+
+### Bug Fixes
+
+* when clear input(not use clear button)  the date value not changed ([39d2c40](https://github.com/mengxiong10/vue2-datepicker/commit/39d2c40))
+
+
+### Features
+
+* add class to .mx-calendar indicating the different panel ([#219](https://github.com/mengxiong10/vue2-datepicker/issues/219)) ([1d0a67b](https://github.com/mengxiong10/vue2-datepicker/commit/1d0a67b))
+* add prop inputAttr ([2381089](https://github.com/mengxiong10/vue2-datepicker/commit/2381089))
+
+
+
+## [2.6.4](https://github.com/mengxiong10/vue2-datepicker/compare/v2.6.3...v2.6.4) (2018-12-19)
+
+
+### Bug Fixes
+
+* when clear input(not use clear button)  the date value not changed ([39d2c40](https://github.com/mengxiong10/vue2-datepicker/commit/39d2c40))
+
+
+
+## [2.6.3](https://github.com/mengxiong10/vue2-datepicker/compare/v2.6.2...v2.6.3) (2018-12-08)
+
+
+### Bug Fixes
+
+* fix unable set value later when range is ture ([#209](https://github.com/mengxiong10/vue2-datepicker/issues/209)) ([97289d1](https://github.com/mengxiong10/vue2-datepicker/commit/97289d1))
+
+
+
+## [2.6.2](https://github.com/mengxiong10/vue2-datepicker/compare/v2.6.1...v2.6.2) (2018-10-30)
+
+
+### Bug Fixes
+
+* `calendar-change` trigger in right time ([b1a5a41](https://github.com/mengxiong10/vue2-datepicker/commit/b1a5a41))
+
+
+### Features
+
+* add `calendar-change` event ([ef9314e](https://github.com/mengxiong10/vue2-datepicker/commit/ef9314e))
+
+
+
+## [2.6.1](https://github.com/mengxiong10/vue2-datepicker/compare/v2.6.0...v2.6.1) (2018-10-17)
+
+
+### Bug Fixes
+
+* prevent popup internal click event from affecting the outside ([de177d8](https://github.com/mengxiong10/vue2-datepicker/commit/de177d8))
+
+
+### Features
+
+* when use script type, install automatic ([a310f59](https://github.com/mengxiong10/vue2-datepicker/commit/a310f59))
+
+
+
+# [2.6.0](https://github.com/mengxiong10/vue2-datepicker/compare/v2.5.0...v2.6.0) (2018-10-11)
+
+
+### Bug Fixes
+
+* fix shortcut color ([ac4aa87](https://github.com/mengxiong10/vue2-datepicker/commit/ac4aa87))
+
+
+### Features
+
+* add prop `appendToBody` ([e26e1f5](https://github.com/mengxiong10/vue2-datepicker/commit/e26e1f5))
+
+
+
 # [2.5.0](https://github.com/mengxiong10/vue2-datepicker/compare/v2.4.3...v2.5.0) (2018-10-05)
+
 
 ### Features
 
 * add panel-change event ([5cdba7b](https://github.com/mengxiong10/vue2-datepicker/commit/5cdba7b))
 
 
-<a name="2.4.3"></a>
+
 ## [2.4.3](https://github.com/mengxiong10/vue2-datepicker/compare/v2.4.0...v2.4.3) (2018-09-28)
 
 
@@ -14,6 +221,7 @@
 
 * fix the selectTime 'am' and 'pm' ([8e475b3](https://github.com/mengxiong10/vue2-datepicker/commit/8e475b3))
 * IE compatibility ([fefed17](https://github.com/mengxiong10/vue2-datepicker/commit/fefed17))
+* timeType default ([38a657b](https://github.com/mengxiong10/vue2-datepicker/commit/38a657b))
 * when select time picker option close popup ([#154](https://github.com/mengxiong10/vue2-datepicker/issues/154)) ([12907ad](https://github.com/mengxiong10/vue2-datepicker/commit/12907ad))
 * year and month disable don't right ([#169](https://github.com/mengxiong10/vue2-datepicker/issues/169)) ([42bc068](https://github.com/mengxiong10/vue2-datepicker/commit/42bc068))
 
@@ -22,53 +230,69 @@
 
 *  emit `clear` event for the clear button ([e0776b6](https://github.com/mengxiong10/vue2-datepicker/commit/e0776b6))
 
-<a name="2.4.0"></a>
+
+
 # [2.4.0](https://github.com/mengxiong10/vue2-datepicker/compare/v2.3.2...v2.4.0) (2018-08-08)
 
 
 ### Features
 
-* add `type` time to show only time picker ([1046731](https://github.com/mengxiong10/vue2-datepicker/commit/1046731))
+* add prop `time` to show only time picker ([1046731](https://github.com/mengxiong10/vue2-datepicker/commit/1046731))
 
-<a name="2.3.2"></a>
+
+
 ## [2.3.2](https://github.com/mengxiong10/vue2-datepicker/compare/v2.2.0...v2.3.2) (2018-08-07)
+
 
 ### Features
 
-* add prop `date-format` to format time header and date tooltip ([3c27647](https://github.com/mengxiong10/vue2-datepicker/commit/3c27647))
+* add prop `date-format` ([3c27647](https://github.com/mengxiong10/vue2-datepicker/commit/3c27647))
 
-<a name="2.2.0"></a>
+
+
 # [2.2.0](https://github.com/mengxiong10/vue2-datepicker/compare/v2.1.0...v2.2.0) (2018-08-06)
+
 
 ### Bug Fixes
 
 * set input autocomplete off ([264458c](https://github.com/mengxiong10/vue2-datepicker/commit/264458c))
 
+
 ### Features
 
 * add event 'change-calendar-year' 'change-calendar-month' ([bc80708](https://github.com/mengxiong10/vue2-datepicker/commit/bc80708))
 
-<a name="2.1.0"></a>
-# [2.1.0]() (2018-07-24)
+
+
+# [2.1.0](https://github.com/mengxiong10/vue2-datepicker/compare/v2.0.0...v2.1.0) (2018-07-24)
+
+
+### Bug Fixes
+
+* fecha.format need a date ([1e54efc](https://github.com/mengxiong10/vue2-datepicker/commit/1e54efc))
+* shortcut date changed to function ([5cf6c72](https://github.com/mengxiong10/vue2-datepicker/commit/5cf6c72))
 
 ### Features
 
-* Add `type` year and month
+* add the slot footer slot-scope to confirm event ([43d9fd7](https://github.com/mengxiong10/vue2-datepicker/commit/43d9fd7))
+* export the fecha to use ([ac958d4](https://github.com/mengxiong10/vue2-datepicker/commit/ac958d4))
 
-<a name="2.0.0"></a>
-# [2.0.0]() (2018-06-16)
+
+
+# [2.0.0](https://github.com/mengxiong10/vue2-datepicker/compare/9df1fe0...v2.0.0) (2018-06-16)
+
 
 ### Features
 
-* Add `clearable` used to show clear icon
-* Add slot `calendar-icon` to custom calendar icon
-* Add slot `header` and `footer` to custom popup area
-* `disabledDays` supports custom functions
+- Add `clearable` used to show clear icon
+- Add slot `calendar-icon` to custom calendar icon
+- Add slot `header` and `footer` to custom popup area
+- `disabledDays` supports custom functions
 
 ### Breaking changes
 
-* Refactored code. This may fail if you hacked the code
-* `format` default value changes from yyyy-MM-dd to YYYY-MM-DD, now the parsing tokens are similar to the moment.js, not supports lowercase yyyy
-* remove `custom-formatter`
-* `editable` default value changes from false to true, now supports range
-* select the year or month panel will not change the value(Because it causes many problems when set the `not-before` or `not-after`)
+- Refactored code. This may fail if you hacked the code
+- `format` default value changes from yyyy-MM-dd to YYYY-MM-DD, now the parsing tokens are similar to the moment.js, not supports lowercase yyyy
+- remove `custom-formatter`
+- `editable` default value changes from false to true, now supports range
+- select the year or month panel will not change the value(Because it causes many problems when set the `not-before` or `not-after`)